requirements.txt
django==3.2.10
django-extensions==3.1.4
djangorestframework==3.11.2
psycopg2-binary